Moves the WebViewFallback into the library
- Moves the WebViewFallbackActivity into the Library
- Prepares the default LauncherActivity to accept a meta-tag
  with the fallback to be used.
- When the fallabck is "webview", launches the WebView fallback.
  Otherwise, launches a custom tab.
