+++ androidbrowserhelper/src/main/java/com/google/androidbrowserhelper/trusted/LauncherActivity.java
@@ -21,8 +21,6 @@ import android.os.Bundle;
 import android.util.Log;
 import android.widget.ImageView;
 
-import com.google.androidbrowserhelper.trusted.splashscreens.PwaWrapperSplashScreenStrategy;
-
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.appcompat.app.AppCompatActivity;
@@ -31,6 +29,8 @@ import androidx.browser.customtabs.CustomTabsIntent;
 import androidx.browser.trusted.TrustedWebActivityIntentBuilder;
 import androidx.browser.trusted.TrustedWebActivityService;
 import androidx.core.content.ContextCompat;
+
+import com.google.androidbrowserhelper.trusted.splashscreens.PwaWrapperSplashScreenStrategy;
 
 /**
  * A convenience class to make using Trusted Web Activities easier. You can extend this class for
@@ -93,6 +93,8 @@ public class LauncherActivity extends AppCompatActivity {
 
     private static final String BROWSER_WAS_LAUNCHED_KEY =
             "android.support.customtabs.trusted.BROWSER_WAS_LAUNCHED_KEY";
+
+    private static final String FALLBACK_TYPE_WEBVIEW = "webview";
 
     /** We only want to show the update prompt once per instance of this application. */
     private static boolean sChromeVersionChecked;
@@ -259,6 +261,16 @@ public class LauncherActivity extends AppCompatActivity {
     }
 
     protected TwaLauncher.FallbackStrategy getFallbackStrategy() {
+        if (FALLBACK_TYPE_WEBVIEW.equalsIgnoreCase(mMetadata.fallbackStrategyType)) {
+            return (context, twaBuilder, providerPackage, completionCallback) -> {
+                Intent intent = WebViewFallbackActivity.createLaunchIntent(LauncherActivity.this,
+                        twaBuilder.getUri(), LauncherActivityMetadata.parse(LauncherActivity.this));
+                context.startActivity(intent);
+                if (completionCallback != null) {
+                    completionCallback.run();
+                }
+            };
+        }
         return TwaLauncher.CCT_FALLBACK_STRATEGY;
     }
