+++ demos/twa-webview-fallback/src/main/AndroidManifest.xml
@@ -13,7 +13,7 @@
             android:name="asset_statements"
             android:resource="@string/asset_statements" />
 
-        <activity android:name=".MyLauncherActivity"
+        <activity android:name="com.google.androidbrowserhelper.trusted.LauncherActivity"
             android:label="@string/app_name">
             <meta-data android:name="android.support.customtabs.trusted.DEFAULT_URL"
                 android:value="https://pwa-directory.appspot.com" />
@@ -41,6 +41,9 @@
             <meta-data android:name="android.support.customtabs.trusted.ADDITIONAL_TRUSTED_ORIGINS"
                 android:resource="@array/additional_trusted_origins" />
 
+            <meta-data android:name="android.support.customtabs.trusted.FALLBACK_STRATEGY"
+                android:value="webview" />
+
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
                 <category android:name="android.intent.category.LAUNCHER" />
@@ -56,7 +59,7 @@
 
         </activity>
 
-        <activity android:name=".WebViewFallbackActivity"
+        <activity android:name="com.google.androidbrowserhelper.trusted.WebViewFallbackActivity"
             android:configChanges="orientation|screenSize" />
 
         <provider
